// src/types/csv.ts
export interface ParsedCsvRow {
  date: string;
  flightNumber: string;
  from: string;
  to: string;
  departureTime: string;
  arrivalTime: string;
  duration: string;
  airline: string;
  aircraft: string;
  registration: string;
  seatNumber: string;
  seatType: string;
  flightClass: string;
  flightReason: string;
  note: string;
  departureId: string;
  arrivalId: string;
  airlineId: string;
  aircraftId: string;
}